requirements.txt
python-chess
pymongo
tqdm
flask
flask-cors
appdirs
